Files added via direct upload
First commit. This entirely new Repo encloses files to support both Radeon RG15 IR Rain Gauge and Calypso ULP Ultrasonic Wind Anemometer
This is work-in-progress.  The main bug with this file is that the PM Board refuses to be discovered on the I2C Bus; the main reason we cannot see readings flow through to the SCK Data Board.